Pisahkan logika `recipeRoute` menjadi beberapa file terpisah dan tambahkan routing baru untuk gambar, komentar, reaksi, serta bookmark.
