+++ server/src/main/kotlin/com/kotlinonly/moprog/core/controller/MetricRoute.kt
